added init_encryption_keys
